Focus the window to ensure cursor can be hidden in fullscreen
"NSCursor.setHiddenUntilMouseMoves" is only effective when the
calling window is the focused window.

When user moves back from another screen or space, the window may
not be focused. As a result, the cursor can't be hidden before the
user clicks the window.

It can be reasonably expected that the user would want to focus on
the window when the mouse moves inside the window and the window is
in fullscreen.

As such, focus the window on mouse movement inside the window when
the window is in fullscreen.

Bug: #1415